Change good grasp depth to 3cm and update inference code for bin inference

+++ infer.py
@@ -14,10 +14,11 @@ from data_utils import (
     BIN_WIDTH_M,
     BIN_LENGTH_M,
     BIN_HEIGHT,
+    WINDOW_SIZE,
 )
 
-WINDOW_SIZE = 50
 BIN_PADDING = 50
+DEBUG_PLOT = True
 
 
 def _tensor_to_numpy_image(t: torch.Tensor) -> np.ndarray:
@@ -195,22 +196,23 @@ def get_patches_from_image_and_depth_maps(
     # Make a copy so the original image isn't modified (OpenCV format, uint8)
     rgb_img_for_viz = rgb_img.copy()
 
-    for center in centers:
-        x, y = center
-        # Top-left and bottom-right of the patch
-        top_left = (x - WINDOW_SIZE // 2, y - WINDOW_SIZE // 2)
-        bottom_right = (x + WINDOW_SIZE // 2 - 1, y + WINDOW_SIZE // 2 - 1)
-        # Draw rectangle (BGR=green)
-        cv2.rectangle(rgb_img_for_viz, top_left, bottom_right, (0, 255, 0), 2)
-        # Draw center (BGR=red)
-        cv2.circle(rgb_img_for_viz, (x, y), 4, (0, 0, 255), -1)
+    if DEBUG_PLOT:
+        for center in centers:
+            x, y = center
+            # Top-left and bottom-right of the patch
+            top_left = (x - WINDOW_SIZE // 2, y - WINDOW_SIZE // 2)
+            bottom_right = (x + WINDOW_SIZE // 2 - 1, y + WINDOW_SIZE // 2 - 1)
+            # Draw rectangle (BGR=green)
+            cv2.rectangle(rgb_img_for_viz, top_left, bottom_right, (0, 255, 0), 2)
+            # Draw center (BGR=red)
+            cv2.circle(rgb_img_for_viz, (x, y), 4, (0, 0, 255), -1)
 
-    # Optional: Show or save the visualization (commented out, for user to enable as needed)
-    cv2.imshow("Patch Visualization", rgb_img_for_viz)
-    cv2.waitKey(0)
-    cv2.destroyAllWindows()
+        # Optional: Show or save the visualization (commented out, for user to enable as needed)
+        cv2.imshow("Patch Visualization", rgb_img_for_viz)
+        cv2.waitKey(0)
+        cv2.destroyAllWindows()
 
-    # return patches_rgb, patches_depth, centers
+    return patches_rgb, patches_depth, centers
 
 
 def infer_on_bin(rgb_img, depth_img, model, device):
@@ -220,25 +222,68 @@ def infer_on_bin(rgb_img, depth_img, model, device):
     depth_img = depth_img[CROP_YMIN:CROP_YMAX, CROP_XMIN:CROP_XMAX]
 
     # Get patches from the image and depth maps
+    patches_rgb, patches_depth, centers = get_patches_from_image_and_depth_maps(
+        rgb_img, depth_img, bin_padding=BIN_PADDING
+    )
 
     # Preprocess the image and depth maps
     transform_rgb = create_transform_rgb()
     transform_depth = create_transform_depth()
-    rgb_img = transform_rgb(rgb_img)
-    depth_img = transform_depth(depth_img)
-    rgb_img = rgb_img.to(device)
-    depth_img = depth_img.to(device)
+    pred_weights = []
+    for rgb_patch, depth_patch, center in zip(patches_rgb, patches_depth, centers):
+        rgb_patch = transform_rgb(rgb_patch)
+        depth_patch = transform_depth(depth_patch)
+        rgb_patch = rgb_patch.to(torch.float32).to(device)
+        depth_patch = depth_patch.to(torch.float32).to(device)
 
-    preds = model(rgb_img, depth_img)
-    preds = preds.squeeze(-1)
-    return preds.item()
+        rgb_patch = rgb_patch.unsqueeze(0)
+        depth_patch = depth_patch.unsqueeze(0)
+
+        pred = model(rgb_patch, depth_patch)
+        pred = pred.squeeze(-1)
+
+        pred_weights.append(pred.item())
+
+    if DEBUG_PLOT:
+        rgb_img_for_viz = cv2.resize(
+            rgb_img.copy(), fx=2, fy=2, interpolation=cv2.INTER_NEAREST, dsize=None
+        )
+        for center, pred_weight in zip(centers, pred_weights):
+            x, y = center
+            x = x * 2
+            y = y * 2
+            # Top-left and bottom-right of the patch
+            # top_left = (x - WINDOW_SIZE // 2, y - WINDOW_SIZE // 2)
+            # bottom_right = (x + WINDOW_SIZE // 2 - 1, y + WINDOW_SIZE // 2 - 1)
+            # Draw rectangle (BGR=green)
+            # cv2.rectangle(rgb_img_for_viz, top_left, bottom_right, (0, 255, 0), 2)
+            # Draw center (BGR=red)
+            # cv2.circle(rgb_img_for_viz, (x, y), 4, (0, 0, 255), -1)
+            cv2.putText(
+                rgb_img_for_viz,
+                f"{pred_weight:.0f}",
+                (x, y),
+                cv2.FONT_HERSHEY_SIMPLEX,
+                0.5,
+                (255, 0, 0),
+                1,
+            )
+
+        cv2.imshow(
+            "Predicted Weights",
+            cv2.resize(
+                rgb_img_for_viz, fx=2, fy=2, interpolation=cv2.INTER_NEAREST, dsize=None
+            ),
+        )
+        cv2.waitKey(0)
+        cv2.destroyAllWindows()
 
 
 def main():
-    model_path = "/home/parth/snaak/projects/granular_grasp/runs/run_test_transforms_1/mass_estimation_model.pth"
+    model_path = "/home/parth/snaak/projects/granular_grasp/runs/train_w50_run_1/mass_estimation_model.pth"
     input_data_dir = "/home/parth/snaak/snaak_data/data_parth"
     output_data_dir = (
-        "/home/parth/snaak/projects/granular_grasp/runs/run_test_1/inference"
+        "/home/parth/snaak/projects/granular_grasp/runs/train_w50_run_1/inference"
     )
     batch_size = 1
     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
@@ -276,5 +321,21 @@ def test_get_patches_from_image_and_depth_maps():
     get_patches_from_image_and_depth_maps(img, depth_map, bin_padding=BIN_PADDING)
 
 
+def test_infer_on_bin():
+
+    model_path = "/home/parth/snaak/projects/granular_grasp/runs/train_w50_run_1/mass_estimation_model.pth"
+    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+    model = MassEstimationModel()
+    model.load_state_dict(torch.load(model_path)["model_state_dict"])
+    model.eval()
+    model.to(device)
+
+    img_path = "/home/parth/snaak/projects/granular_grasp/rgb_image.jpg"
+    img = cv2.imread(img_path)
+    depth_map = np.zeros((img.shape[0], img.shape[1]))
+    depth_map[img.shape[0] // 2, img.shape[1] // 2] = 100
+    infer_on_bin(img, depth_map, model, device)
+
+
 if __name__ == "__main__":
-    test_get_patches_from_image_and_depth_maps()
+    test_infer_on_bin()
